requirements.txt
pillow==10.3.0
PyAutoGUI==0.9.54